pyproject.toml
[project]
name = "systemd-web"
version = "0.0.1"
description = "A super simple web control for start/stop/restart/status of systemd services"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]
